Refactor Content
