+++ VisualStudio/Content/Reworks/FocusedConvergenceRework.cs
